+++ client/src/pages/Dashboard/index.js
@@ -3,6 +3,7 @@ import { Layout } from 'antd';
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import AdminMenu from '../../components/Menu';
+import Navbar from '../../components/Navbar';
 import Rooms from '../../components/Table/Rooms';
 import Users from '../../components/Table/Users';
 import './style.css';
@@ -12,10 +13,11 @@ const { Content, Sider } = Layout;
 function Dashboard() {
   return (
     <Layout>
-      <Sider>
-        <AdminMenu />
-      </Sider>
+      <Navbar />
       <Layout>
+        <Sider>
+          <AdminMenu />
+        </Sider>
         <Content
           style={{
             padding: 24,
